Move svg components to new folder
Moved from /src/assets to /src/components/svg
