+++ src/components/header/Header.tsx
@@ -1,7 +1,7 @@
 import { useSelector } from "react-redux";
 
-import AddIcon from "~/assets/AddIcon";
-import Logo from "~/assets/Logo";
+import AddIcon from "~/components/svg/AddIcon";
+import Logo from "~/components/svg/Logo";
 import Button from "~/components/ui/Button";
 import Submenu from "~/components/ui/Submenu";
 import type { RootState } from "~/store/store";
@@ -14,7 +14,7 @@ const Header: React.FC = () => {
   );
 
   return (
-    <header className="flex items-center h-16 bg-white dark:bg-gunmetal-800 md:h-20 lg:h-24">
+    <header className="flex h-16 items-center bg-white dark:bg-gunmetal-800 md:h-20 lg:h-24">
       <div
         className={`hidden h-full items-center border-r border-indigo px-4 dark:border-gunmetal-700 md:flex md:px-6 ${
           showSidebar && "w-64"
@@ -27,7 +27,7 @@ const Header: React.FC = () => {
           <Logo className="hidden md:block" dark />
         </span>
       </div>
-      <div className="flex items-center justify-between flex-1 w-full pl-4 md:pl-6 md:pr-2">
+      <div className="flex w-full flex-1 items-center justify-between pl-4 md:pl-6 md:pr-2">
         {/* <SidebarMobile
           // TODO only needs names not all board data?
           boards={props.boards}
