+++ src/components/svg/BoardIcon.tsx
