+++ src/components/svg/CloseSidebarIcon.tsx
