+++ src/components/svg/OpenSidebarIcon.tsx
