+++ src/components/svg/SignoutIcon.tsx
